backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.10.1
sqlparse==0.5.5
typing_extensions==4.15.0
